requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.103.2
greenlet==3.0.3
h11==0.14.0
idna==3.7
importlib-metadata==6.7.0
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0
